requirements.txt
pvporcupine
sounddevice
openai-whisper
torch
pyautogui
pyaudio
pystray
Pillow
pyperclip
pygetwindow